chore: move Apollo client config in its own file
